requirements.txt
newsapi-python
google-generativeai
schedule
python-dotenv
